fix(register): B2B-4861 Buyer Portal: Add null check for stateRequired flag and remove FF
