+++ apps/storefront/src/pages/Registered/config.ts
@@ -47,15 +47,6 @@ export interface State {
   id?: string;
 }
 
-export function getIsStateRequired(
-  country: { stateRequired?: boolean } | undefined,
-  stateList: { length: number } | null | undefined,
-): boolean {
-  return typeof country?.stateRequired === 'boolean'
-    ? country.stateRequired
-    : (stateList?.length ?? 0) > 0;
-}
-
 type EmailError = {
   [k: number]: string;
 };
