+++ apps/storefront/src/utils/b2bGetIsStateRequired.ts
@@ -1,0 +1,8 @@
+export function getIsStateRequired(
+  country: { stateRequired?: boolean } | undefined,
+  stateList: { length: number } | null | undefined,
+): boolean {
+  return typeof country?.stateRequired === 'boolean'
+    ? country.stateRequired
+    : (stateList?.length ?? 0) > 0;
+}
